Add : Mission 1 - 사용자 입력 데이터 저장 기능 구현
