+++ src/pages/cheoljinPark/Login/LoginForm.scss
@@ -1,0 +1,33 @@
+.login-form {
+  .loginInput {
+    display: block;
+    width: 100%;
+    height: 70px;
+    border: 2px solid #ccc;
+    border-radius: 5px;
+    margin-bottom: 10px;
+    font-size: 17px;
+    padding: 15px;
+  }
+
+  .login-btn {
+    margin: 15px 0;
+    width: 100%;
+    height: 50px;
+    border-radius: 5px;
+    box-shadow: none;
+    background-color: #497bde;
+    border: 0 solid black;
+    font-size: 23px;
+    color: #fff;
+    opacity: 0.5;
+  }
+
+  .login-btn:hover {
+    cursor: pointer;
+  }
+
+  .actived {
+    opacity: 1;
+  }
+}
